Cleaned up search page

+++ server/queryEvents.js
@@ -290,17 +290,19 @@ module.exports = queryEvents = () => {
         .then(apiResponse => {
           const channels = apiResponse.data.channels;
 
-          channels.forEach(channel => {
-            axios.delete(
-              `https://api-${appId}.sendbird.com/v3/group_channels/${channel.channel_url}`,
-              {
-                headers: {
-                  'Content-Type': 'application/json; charset=utf8',
-                  'Api-Token': SB_API_TOKEN,
-                },
-              }
-            );
-          });
+          if (channels && channels.length > 0) {
+            channels.forEach(channel => {
+              axios.delete(
+                `https://api-${appId}.sendbird.com/v3/group_channels/${channel.channel_url}`,
+                {
+                  headers: {
+                    'Content-Type': 'application/json; charset=utf8',
+                    'Api-Token': SB_API_TOKEN,
+                  },
+                }
+              );
+            });
+          }
         })
         .catch(e => log.error(e.data));
     }
